Clean up add-new-deck: colocate function, use mutation
This cleanup moves the submit function out of `posters.ts` and colocates
it alongside the components/routes that use it, as `add-new-deck.ts`.

I enjoy this pattern, as the functions are only being used once, but they
could be re-used elsewhere, and regardless it's still nice to give them their
own space, and to be able to use typescript.

It also separates the page into server side page.jsx and the 'use client'
component in form.jsx.
